requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cloud-sql-python-connector==1.18.2
colorama==0.4.6
comm==0.2.2
cryptography==45.0.4
debugpy==1.8.14
decorator==5.2.1
dnspython==2.7.0
executing==2.2.0
frozenlist==1.7.0
google-auth==2.40.3
greenlet==3.2.3
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kagglehub==0.3.12
matplotlib-inline==0.1.7
multidict==6.4.4
nest-asyncio==1.6.0
numpy==2.3.0
packaging==25.0
pandas==2.3.0
parso==0.8.4
pg8000==1.31.2
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.32.4
rsa==4.9.1
scramp==1.4.5
six==1.17.0
SQLAlchemy==2.0.41
stack-data==0.6.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
yarl==1.20.1
